+++ changelog.md
@@ -14,6 +14,7 @@ compromise uses semver, and pushes to npm frequently
 - support whitespace before sentence boundary
 - improved QuestionWord tagging, some `.questions()` without a question-mark
 - phrasalVerb conjugation
+- new #Activity tag for Gerunds as nouns 'walking is fun'
 <!-- * include 'the #TitleCase' matches in .topics() -->
 
 ### v10